+++ test/unit/fuzz-tools.test.js
@@ -242,6 +242,7 @@ import * as renalpulmV249 from '../../lib/renalpulm-v249.js';
 import * as obgynV250 from '../../lib/obgyn-v250.js';
 import * as cardiometabV251 from '../../lib/cardiometab-v251.js';
 import * as orthospineV252 from '../../lib/orthospine-v252.js';
+import * as radmeasureV253 from '../../lib/radmeasure-v253.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -428,6 +429,7 @@ const MODULES = {
   'obgyn-v250.js': obgynV250,
   'cardiometab-v251.js': cardiometabV251,
   'orthospine-v252.js': orthospineV252,
+  'radmeasure-v253.js': radmeasureV253,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];